build: update llvm tag to 14212c96
Summary of changes:
 - Updated MHLO commit to build with LLVM commit hash 14212c96
 - Re-enabled two failing TOSA tests, thus resolving Github Issue #1231
 - Replace deprecated hasValue() with has_value()
   (https://reviews.llvm.org/D131349)
 - Fixed warnings from comparison between signed and unsigned values

+++ lib/Dialect/Torch/Transforms/DecomposeComplexOps.cpp
@@ -871,7 +871,7 @@ public:
       }
 
       newSizes.reserve(rank - end + start);
-      for (size_t k = 0; k < start; ++k) {
+      for (int64_t k = 0; k < start; ++k) {
         Value dim =
             rewriter.create<ConstantIntOp>(loc, rewriter.getI64IntegerAttr(k));
         newSizes.push_back(
@@ -880,7 +880,7 @@ public:
       Value flattenDimSize =
           rewriter.create<ConstantIntOp>(loc, rewriter.getI64IntegerAttr(-1));
       newSizes.push_back(flattenDimSize);
-      for (size_t k = end + 1; k < rank; ++k) {
+      for (int64_t k = end + 1; k < rank; ++k) {
         Value dim =
             rewriter.create<ConstantIntOp>(loc, rewriter.getI64IntegerAttr(k));
         newSizes.push_back(
